+++ src/db/products/all-products.json
@@ -1,0 +1,25 @@
+[{
+        "id": 19112831,
+        "sku": 1120001,
+        "name": "Пицца Пепперони с томатами",
+        "description": "Пицца пепперони - очень популярный вид пиццы как в Европе, так и в Америке. Она получила свое название благодаря основному ингредиенту солями пепперони, который и отличает ее от всех остальных видов пицц. Он имеет острый вкус и при запекании покрывается хрустящей жирной корочкой.",
+        "price": "100",
+        "currency": "UAN",
+        "creatorId": 1,
+        "created": "21-08-2018",
+        "modified": "21-08-2018",
+        "categories": ["pizza"]
+    },
+    {
+        "id": 19112832,
+        "sku": 1120002,
+        "name": "Пицца Маргарита",
+        "description": "Пожалуй, самая популярная в мире, даже меню любой пиццерии начинается, как правило, именно с неё. Состав этой пиццы необычайно прост, её основные ингредиенты: сыр моцарелла, спелые помидоры и листья свежего базилика, которые придают ей неповторимый вкус и аромат.",
+        "price": "80",
+        "currency": "UAN",
+        "creatorId": 1,
+        "created": "21-08-2018",
+        "modified": "21-08-2018",
+        "categories": ["pizza"]
+    }
+]
