+++ src/routers/main/main.js
@@ -1,7 +1,7 @@
     const mainRoute = (request, response) => {
 
         response.writeHead(200, { "Content-Type": "text/html" });
-        response.write("<h1>Hello world!</h1>");
+        response.write("<h1>It`s main page!</h1>");
         response.end();
 
     };
