+++ src/routers/router.js
@@ -1,11 +1,11 @@
 const mainRoute = require('./main/main');
-const motocycleRoute = require('./motocycle/motocycle');
-const signUpRoute = require('./users/sign-up-route');
+const products = require('./products/products');
+const signUp = require('./users/sign-up');
 
 const router = {
-  '/signup': signUpRoute,
-  '/motocycle': motocycleRoute,
-  default: mainRoute
+    '/signup': signUp,
+    '/products': products,
+    default: mainRoute
 };
 
 module.exports = router;
